Wordpress Plugin
This was made purely for learning more about Wordpress plugins.  Input your location on the backend of wordpress and this widget will output a sentence telling the viewer what the weather is in that location. This widget is intended for the sidebar. It is a minimal, text only widget inspired by Location Weather By ShapedPlugin. I manipulated the pre-existing functionality from their widget to make my own.