Creating our new objects
First of all, change your existing Ball() constructor so that it becomes a Shape() constructor and add a new Ball() constructor:

The Shape() constructor should define the x, y, velX, and velY properties in the same way as the Ball() constructor did originally, but not the color and size properties.
It should also define a new property called exists, which is used to track whether the balls exist in the program (have not been eaten by the evil circle). This should be a boolean (true/false).
The Ball() constructor should inherit the x, y, velX, velY, and exists properties from the Shape() constructor.
It should also define a color and a size property, like the original Ball() constructor did.
Remember to set the Ball() constructor's prototype and constructor appropriately.
The ball draw(), update(), and collisionDetect() method definitions should be able to stay exactly the same as they were before.

You also need to add a new parameter to the new Ball() ( ... ) constructor call — the exists parameter should be the 5th parameter, and should be given a value of true.

At this point, try reloading the code — it should work just the same as it did before, with our redesigned objects.