define EvilCircle constructor
Now it's time to meet the bad guy — the EvilCircle()! Our game is only going to involve one evil circle, but we are still going to define it using a constructor that inherits from Shape() to give you some practice. You might want to add another circle to the app later on that can be controlled by another player, or have several computer-controlled evil circles. You're probably not going to take over the world with a single evil circle, but it will do for this assessment.

The EvilCircle() constructor should inherit x, y, velX, velY, and exists from Shape(), but velX and velY should always equal 20.

You should do this something like Shape.call(this, x, y, 20, 20, exists);

It should also define its own properties, as follows:

color — 'white'
size — 10
Again, remember to define your inherited properties as parameters in the constructor, and set the prototype and constructor properties correctly.