Implementing the score counter
To implement the score counter, follow the following steps:

In your HTML file, add a <p> element just below the <h1> element containing the text "Ball count: ".
In your CSS file, add the following rule at the bottom:
p {
  position: absolute;
  margin: 0;
  top: 35px;
  right: 5px;
  color: #aaa;
}
In your JavaScript, make the following updates:
Create a variable that stores a reference to the paragraph.
Keep a count of the number of balls on screen in some way.
Increment the count and display the updated number of balls each time a ball is added to the scene.
Decrement the count and display the updated number of balls each time the evil circle eats a ball (causes it not to exist).